chore: moving stream_map under the cluster::controller mod
